Fix up test logic as per my understanding.
When a query is created and enabled at the same time, the query will
run until completion.

However, it seems that from the test (a_completed_projection) that
another "Enable" message gets sent and with that, the projection's
status will be updated (given that the query was not enabled, more on
that later). e.g. Completed -> Stopped

The same logic applies for the second test (a_failed_projection), where
my assumption would be that if the projection's source is "bad", the
status of the projection would be "faulted" and not running.

There was a change to make projection operations idempotent. This
change broke these 2 tests only.

https://github.com/EventStore/EventStore/commit/445f9ef01aed52ab48eb375947771248b1ce6675